chore: split medians into two files
