+++ state-chain/pallets/cf-elections/src/electoral_systems.rs
@@ -1,4 +1,5 @@
 pub mod blockchain;
 pub mod change;
 pub mod composite;
-pub mod median;
+pub mod monotonic_median;
+pub mod unsafe_median;
